Move form list component from BMI profile folder to general "formly" directory for reusability
