+++ frontend/src/app/formly/types/repeat-distribution-detail-list/repeat-distribution-detail-list.component.scss
